Organise code for better readablity
